fix: faster solution
Correct the logic behind accessing the j-th element and i-th element to compare them in main by using the connection between pointers instead of travelling the list from start to the desired element. Also, modify the List constructor so that the nodes are not in charge of their own deletion (they are not necessarily in a list). To do next should be correcting the header files since and implementing the new ModifiedList, since all the relevant changes were made in the single-file solution to the problem.