Assignment 2: Binary tree traversal
Multiple attempts to solve this problem are created. The Python implementation is as optimized as possible, yet it exceeds the time limit in the fifth case. In contrast, the second attempt in C++ is done using the essential idea for the problem (partitioning the inorder according to the roots in the preorder). However, this one too, due to creating multiple arrays, is too inefficient. Finally, the binary_tree_traversal.cpp file contains the optimized solution: making a map to access the elements in inorder (to make the search for the root in inorder constant), tracking the indices of the arrays instead of making new ones (just like the Python one), and building one array instead of adding them together to create a new one (which is one of the reasons why there is the need to switch to C++).